# Add native ARM64 popcount for sizeof(T) above 1 to TensorPrimitives